Now GlobalVulkanContext vkContext (for now it's just vkContext.device) is initialized in VulkanDevice class. The vkContext variable will be used (not for all these classes have yet been adapted to this paradigm). For future improvement we can modify the GlobalVulkanContext class (struct will be changed to class) to more precisely manipulate the Vulkan context. All methods of the VulkanRenderBackend class used for the main rendering loop have been added.

+++ Engine/Nutckracker/Renderer/Buffer.hpp
@@ -69,7 +69,7 @@ namespace NK {
 		std::vector<Vertex> m_Vertices_{};
 		std::vector<uint32_t> m_Indices_{};
 		
-		void buildModel(float* vertices, uint32_t v_size, float* indices, uint32_t i_size = 0);
+		void buildModel(float* vertices, uint32_t v_size, uint32_t* indices, uint32_t i_size = 0);
 		void loadModel(const std::string& filepath) { /*empty inplementation*/ }
 	};
